+++ examples/extlb_fgtasg_intlb/main.tf
@@ -56,15 +56,24 @@ module "fortigate_scaleset" {
   storage_account_name          = try(each.value.storage_account_name, null)
   storage_account_creation_flag = try(each.value.storage_account_creation_flag, true)
   vmss_name                     = try(each.value.vmss_name, "fortigate-scaleset")
-  image_version                 = try(each.value.image_version, "7.2.8")
-  license_type                  = try(each.value.license_type, "byol")
-  # image_sku                     = format("fortinet_fg-vm%s%s", each.value.license_type == "payg" ? "_payg_2023" : "", try(each.value.architecture, "") == "Arm64" ? "_arm64" : "")
+
+  image_version = can(regex("^\\d+\\.\\d+$", each.value.image_version)) ? "latest" : each.value.image_version
+
+  license_type = try(each.value.license_type, "byol")
 
   image_sku = format(
-    "fortinet_fg-vm%s%s%s",
-    each.value.license_type == "payg" ? "_payg_2023" : "",
-    try(each.value.architecture, "") == "Arm64" ? "_arm64" : "",
-    contains(["7.6.1", "7.6.2", "7.6.3", "7.6.4"], each.value.image_version) ? "_g2" : ""
+    "fortinet_fg-vm_%s_%s%s%s",
+    try(each.value.license_type, "byol"),
+    replace(
+      regex(
+        "^\\d+\\.\\d+",
+        each.value.image_version
+      ),
+      ".",
+      ""
+    ),
+    lower(try(each.value.gen_type, "standard")) == "g2" ? "_g2" : "",
+    lower(try(each.value.architecture, "")) == "arm64" ? "_arm64" : ""
   )
 
   application_insights_id = try(each.value.application_insights_id, null)
@@ -79,7 +88,7 @@ module "fortigate_scaleset" {
   ]
 
   fortigate_username            = try(each.value.fortigate_username, "fgtadmin")
-  fortigate_password            = try(each.value.fortigate_password, random_password.random_fgt_password)
+  fortigate_password            = try(each.value.fortigate_password, random_password.random_fgt_password.result)
   fortigate_license_folder_path = try("${path.cwd}/${each.value.fortigate_license_folder_path}", "./licenses")
   fortigate_autoscale_psksecret = random_password.psksecret.result
   fortigate_custom_config = templatefile(try(each.value.fortigate_custom_config_file_path, "${path.module}/fortigate_custom_config.conf"), {
@@ -92,6 +101,7 @@ module "fortigate_scaleset" {
   enable_accelerated_networking = try(each.value.enable_accelerated_networking, true)
   data_type                     = try(each.value.data_type, "custom_data")
   autoscale_notification_emails = try(each.value.autoscale_notification_emails, [])
+  fmg_integration               = try(each.value.fmg_integration, null)
   min_count                     = try(each.value.min_count, 1)
   max_count                     = try(each.value.max_count, 1)
   default_count                 = try(each.value.default_count, 1)