+++ modules/fortigate/scaleset/main.tf
@@ -1,11 +1,11 @@
 locals {
-  handle_scale_event_endpoint = format(
+  handle_scale_event_endpoint = var.fmg_integration != null ? "" : format(
     "https://%s/api/handle_auto_scale_events?code=%s",
-    azurerm_linux_function_app.function_app.default_hostname,
-    data.azurerm_function_app_host_keys.function_app_keys.default_function_key
+    azurerm_linux_function_app.function_app[0].default_hostname,
+    data.azurerm_function_app_host_keys.function_app_keys[0].default_function_key
   )
   fortinet_sku_to_versions_map = jsondecode(data.local_file.fortinet_sku_to_versions_map_file.content)
-  image_version_valid          = contains(local.fortinet_sku_to_versions_map[var.image_sku], var.image_version)
+  image_version_valid          = can(local.fortinet_sku_to_versions_map[var.image_sku]) && contains(local.fortinet_sku_to_versions_map[var.image_sku], var.image_version)
 
   # Interface details
   public_interface_names                 = [for nic in var.network_interfaces : nic.name if try(nic.create_pip, false)]
@@ -23,6 +23,7 @@ locals {
     custom_config                        = var.fortigate_custom_config
     license_type                         = var.license_type
     fmg_integration                      = var.fmg_integration
+    fortigate_autoscale_psksecret        = var.fortigate_autoscale_psksecret
   })
 }
 
@@ -87,7 +88,7 @@ resource "azurerm_linux_virtual_machine_scale_set" "vmss" {
 
   source_image_reference {
     publisher = "fortinet"
-    offer     = "fortinet_fortigate-vm_v5"
+    offer     = "fortinet_fortigate-vm"
     sku       = var.image_sku
     version   = var.image_version
   }
@@ -95,7 +96,10 @@ resource "azurerm_linux_virtual_machine_scale_set" "vmss" {
   plan {
     name      = var.image_sku
     publisher = "fortinet"
-    product   = "fortinet_fortigate-vm_v5"
+    product   = "fortinet_fortigate-vm"
+  }
+
+  boot_diagnostics {
   }
 
   depends_on = [null_resource.validate_image_version]
@@ -144,8 +148,11 @@ resource "azurerm_monitor_autoscale_setting" "autoscale_setting" {
     email {
       custom_emails = var.autoscale_notification_emails
     }
-    webhook {
-      service_uri = local.handle_scale_event_endpoint
+    dynamic "webhook" {
+      for_each = local.handle_scale_event_endpoint != "" ? [1] : []
+      content {
+        service_uri = local.handle_scale_event_endpoint
+      }
     }
   }
 
@@ -153,25 +160,26 @@ resource "azurerm_monitor_autoscale_setting" "autoscale_setting" {
 }
 
 resource "azurerm_log_analytics_workspace" "log_analytics_workspace" {
+  count               = var.fmg_integration == null ? 1 : 0
   name                = "${var.vmss_name}-law"
   location            = var.location
   resource_group_name = var.resource_group_name
   sku                 = "PerGB2018"
   retention_in_days   = 30
-
-  tags = var.tags
+  tags                = var.tags
 }
 
 resource "azurerm_application_insights" "app_insights" {
+  count               = var.fmg_integration == null ? 1 : 0
   name                = "${var.vmss_name}-appinsights"
   location            = var.location
   resource_group_name = var.resource_group_name
-  workspace_id        = azurerm_log_analytics_workspace.log_analytics_workspace.id
+  workspace_id        = azurerm_log_analytics_workspace.log_analytics_workspace[0].id
   application_type    = "web"
 }
 
 resource "random_string" "random_storage_account_name" {
-  count   = var.storage_account_creation_flag ? 1 : 0
+  count   = var.storage_account_creation_flag && var.fmg_integration == null ? 1 : 0
   length  = 10
   upper   = false
   special = false
@@ -179,7 +187,7 @@ resource "random_string" "random_storage_account_name" {
 
 # Create a new storage account if none is provided
 resource "azurerm_storage_account" "new_account" {
-  count                    = var.storage_account_creation_flag ? 1 : 0
+  count                    = var.storage_account_creation_flag && var.fmg_integration == null ? 1 : 0
   name                     = "fgtvmss${random_string.random_storage_account_name[0].result}"
   resource_group_name      = var.resource_group_name
   location                 = var.location
@@ -188,13 +196,15 @@ resource "azurerm_storage_account" "new_account" {
 }
 
 data "azurerm_storage_account" "account" {
+  count               = var.fmg_integration == null ? 1 : 0
   resource_group_name = var.resource_group_name
   name                = var.storage_account_creation_flag ? azurerm_storage_account.new_account[0].name : var.storage_account_name
   depends_on          = [azurerm_storage_account.new_account]
 }
 
 data "azurerm_storage_account_sas" "account_sas" {
-  connection_string = data.azurerm_storage_account.account.primary_connection_string
+  count             = var.fmg_integration == null ? 1 : 0
+  connection_string = data.azurerm_storage_account.account[0].primary_connection_string
   https_only        = true
   resource_types {
     service   = true
@@ -229,6 +239,7 @@ data "azurerm_storage_account_sas" "account_sas" {
 
 # App Service: Service Plan.
 resource "azurerm_service_plan" "plan" {
+  count               = var.fmg_integration == null ? 1 : 0
   name                = "${var.vmss_name}-appserviceplan"
   resource_group_name = var.resource_group_name
   location            = var.location
@@ -237,18 +248,20 @@ resource "azurerm_service_plan" "plan" {
 }
 
 resource "random_string" "random_function_app_name" {
+  count   = var.fmg_integration == null ? 1 : 0
   length  = 6
   upper   = false
   special = false
 }
 
 resource "azurerm_linux_function_app" "function_app" {
-  name                       = "${var.vmss_name}-func-${random_string.random_function_app_name.result}"
+  count                      = var.fmg_integration == null ? 1 : 0
+  name                       = "${var.vmss_name}-func-${random_string.random_function_app_name[0].result}"
   resource_group_name        = var.resource_group_name
   location                   = var.location
-  service_plan_id            = azurerm_service_plan.plan.id
-  storage_account_name       = data.azurerm_storage_account.account.name
-  storage_account_access_key = data.azurerm_storage_account.account.primary_access_key
+  service_plan_id            = azurerm_service_plan.plan[0].id
+  storage_account_name       = data.azurerm_storage_account.account[0].name
+  storage_account_access_key = data.azurerm_storage_account.account[0].primary_access_key
 
   site_config {
     application_stack {
@@ -261,17 +274,17 @@ resource "azurerm_linux_function_app" "function_app" {
   }
 
   app_settings = {
-    "AzureWebJobsStorage"                   = data.azurerm_storage_account.account.primary_blob_connection_string
-    "WEBSITE_RUN_FROM_PACKAGE"              = "https://${data.azurerm_storage_account.account.name}.blob.core.windows.net/${azurerm_storage_container.container.name}/${azurerm_storage_blob.function_blob.name}?${data.azurerm_storage_account_sas.account_sas.sas}"
-    "APPLICATIONINSIGHTS_CONNECTION_STRING" = azurerm_application_insights.app_insights.connection_string
+    "AzureWebJobsStorage"                   = data.azurerm_storage_account.account[0].primary_blob_connection_string
+    "WEBSITE_RUN_FROM_PACKAGE"              = "https://${data.azurerm_storage_account.account[0].name}.blob.core.windows.net/${azurerm_storage_container.container[0].name}/${azurerm_storage_blob.function_blob[0].name}?${data.azurerm_storage_account_sas.account_sas[0].sas}"
+    "APPLICATIONINSIGHTS_CONNECTION_STRING" = azurerm_application_insights.app_insights[0].connection_string
     "FUNCTIONS_WORKER_RUNTIME"              = "python"
     "FUNCTIONS_EXTENSION_VERSION"           = "~4"
     "AZURE_SUBSCRIPTION_ID"                 = var.azure_subscription_id
     "RESOURCE_GROUP_NAME"                   = var.resource_group_name
     "VMSS_NAME"                             = var.vmss_name
-    "STORAGE_CONTAINER_NAME"                = azurerm_storage_container.container.name
-    "STORAGE_ACCOUNT_NAME"                  = data.azurerm_storage_account.account.name
-    "STORAGE_SAS_CONFIG"                    = data.azurerm_storage_account_sas.account_sas.sas
+    "STORAGE_CONTAINER_NAME"                = azurerm_storage_container.container[0].name
+    "STORAGE_ACCOUNT_NAME"                  = data.azurerm_storage_account.account[0].name
+    "STORAGE_SAS_CONFIG"                    = data.azurerm_storage_account_sas.account_sas[0].sas
     "PRIVATE_INTERFACE_NAME"                = length(local.private_interface_names) == 1 ? local.private_interface_names[0] : ""
     "GWLB_FRONTEND_IP_ADDRESS"              = length(local.lb_frontend_ip_addresses) == 1 ? local.lb_frontend_ip_addresses[0] : ""
     "FORTIGATE_INSTANCE_USER_NAME"          = var.fortigate_username
@@ -294,7 +307,8 @@ resource "azurerm_linux_function_app" "function_app" {
 }
 
 data "azurerm_function_app_host_keys" "function_app_keys" {
-  name                = azurerm_linux_function_app.function_app.name
+  count               = var.fmg_integration == null ? 1 : 0
+  name                = azurerm_linux_function_app.function_app[0].name
   resource_group_name = var.resource_group_name
 
   depends_on = [azurerm_linux_function_app.function_app]
@@ -302,23 +316,26 @@ data "azurerm_function_app_host_keys" "function_app_keys" {
 
 # Assign Reader role to Function App's Managed Identity
 resource "azurerm_role_assignment" "role_assignment" {
-  principal_id         = azurerm_linux_function_app.function_app.identity[0].principal_id
+  count                = var.fmg_integration == null ? 1 : 0
+  principal_id         = azurerm_linux_function_app.function_app[0].identity[0].principal_id
   role_definition_name = "Reader"
   scope                = azurerm_linux_virtual_machine_scale_set.vmss.id
 }
 
 # Define a storage container for function code
 resource "azurerm_storage_container" "container" {
+  count                 = var.fmg_integration == null ? 1 : 0
   name                  = "function-code"
-  storage_account_id    = data.azurerm_storage_account.account.id
+  storage_account_id    = data.azurerm_storage_account.account[0].id
   container_access_type = "private"
 }
 
 # Upload the Function App package to the storage account
 resource "azurerm_storage_blob" "function_blob" {
+  count                  = var.fmg_integration == null ? 1 : 0
   name                   = "functionapp.zip"
-  storage_account_name   = data.azurerm_storage_account.account.name
-  storage_container_name = azurerm_storage_container.container.name
+  storage_account_name   = data.azurerm_storage_account.account[0].name
+  storage_container_name = azurerm_storage_container.container[0].name
   type                   = "Block"
   source                 = "${path.module}/function_app.zip"
 
@@ -329,21 +346,22 @@ resource "azurerm_storage_blob" "function_blob" {
 
 # Upload all file type licenses to the container
 resource "azurerm_storage_blob" "fortigate_licenses" {
-  for_each               = var.fortigate_license_folder_path != null ? fileset(var.fortigate_license_folder_path, "*.lic") : []
+  for_each               = var.fmg_integration == null ? (var.fortigate_license_folder_path != null ? fileset(var.fortigate_license_folder_path, "*.lic") : toset([])) : toset([])
   name                   = "licenses/${each.value}"
-  storage_account_name   = data.azurerm_storage_account.account.name
-  storage_container_name = azurerm_storage_container.container.name
+  storage_account_name   = data.azurerm_storage_account.account[0].name
+  storage_container_name = azurerm_storage_container.container[0].name
   type                   = "Block"
   source                 = "${var.fortigate_license_folder_path}/${each.value}"
 }
 
 data "http" "function_health_check" {
+  count              = var.fmg_integration == null ? 1 : 0
   url                = local.handle_scale_event_endpoint
   method             = "POST"
   request_body       = "{\"reason\": \"TFE health check\"}"
   request_timeout_ms = 3 * 60 * 1000
   retry {
-    attempts = 6
+    attempts = 8
   }
 
   lifecycle {
@@ -368,7 +386,7 @@ resource "null_resource" "validate_image_version" {
   provisioner "local-exec" {
     command = <<EOT
     if ! ${local.image_version_valid}; then
-      echo "Error: Fortinet image version ${var.image_version} not found, please check with the command  az vm image list -o table --all --publisher fortinet --offer fortinet_fortigate-vm_v5 to list all the available products!"
+      echo "Error: Fortinet image version ${var.image_version} not found, please check with the command  az vm image list -o table --all --publisher fortinet --offer fortinet_fortigate-vm to list all the available products!"
       exit 1
     fi
     EOT